Update helpers.php
By default laravel symlinks `public/storage` to `storage/app/public`. This means when trying to retrieve files from `public/storage` using the `asset()` helper you need to prefix your paths with `storage/`. The PR adds a public `public_storage()` which returns the the same as the `asset()` but automatically prefixed with `storage/`. This works well with 5.3 news `->store()` method for easy file uploads.