Proposal: create tracing on Eloquent Models
Similar to what Laravel has currently with timestamps ('created_at' /
'updated_at') on models, I propose to the creation o a (very) basic user
action tracing system through 'created_by' (for the author) and
'updated_by' (for the last editor) attributes on models.

I've had this need on a project I'm working on, and thought that maybe
it could be useful to someone else.

TODO: add methods on Model class to retrieve the author instance and
last editor instance. I have to dig a bit on this as the user provider
depends on the provider that the user configures. Just wanted to check
with you if this is valid or not before I do that.