Allow Memcached to cache for more than 30 days.
This commit will allow Memcached to cache entries for more than
30 days. Every expiration time in Memcached will be converted to
seconds, unless they exceed 30 days. Then, instead of seconds, a
unix timestamp is expected.

If not adjusted, this will most likely result in an immediate
expiration of the entry stored. An example would be:

1. The user stores an entry for 40 days, supplying 57600 minutes.
2. The minutes will get converted to 3456000 seconds.
3. Since the timespan exceeds 30 days, Memcached now expects a
   unix timestamp.
4. The unix timestamp would be "1970-02-10 00:00:00", meaning the
   entry stored is already expired upon store.

To avoid this, every value greater than 30 days, will automatically
be converted to the correct unix timestamp.

This commit will break programs that already are aware of that
behaviour and used a workaround for it. However, this will make
sure that the Memcached implementation is compatible to the other
caching providers and will avoid unexpected behaviour.