Allow the Notification to specify the "Key" it should use to broadcast
I don't like to expose the User.id in my public JavaScript files.
I use Hashid's to hash the user id but then i came across the issue that my notification object could not specify the key it should use to broadcast. This change allows the Notification to specify a different key or else it just uses the getKey method on the Notifiable object.