Added a filter to get only the file inputs called with this->attach(input, value)
Note: This only happens when running tests.

If you are submitting a form that has input file fields and don't attach any file on those inputs
it will send those as normal inputs with a value of null causing a validation exception if any of
those fields have a validation rule of any kind (except required).

This happens because on how validation of null fields are handled now in 5.3.

If you don't expect those fields to be nullable, you'll have to add the 'nullable' validation rule to
make your test pass.

I think there is no use in sending empty file fields as normal inputs with a value of null.

This filter prevents sending empty input file fields as normal inputs with a null value.

File inputs in an html form...
```
    // within an html form in a view
    <input type="file" name="front_photo">
    <input type="file" name="back_photo">
    <input type="file" name="left_photo">
    <input type="file" name="right_photo">
    <input type="submit" value="submit">
```

Test submitting the form...
```
    // within a test class method...
    $this->attach('path-to-file.ext', 'front_photo');

    $this->press('submit');
```

Inspecting the input whitin the controller...
```
    // without the filter i get:
    [
        back_photo => null
        left_photo => null
        right_photo => null
    ]

    // with the filter i get:
    [
    ]
```