Added ability to specify middleware to skip during tests
In response to the issue of CSRF middleware which rarely works in tests, and because I don't want to entirely disable
*all* middleware while testing, this is a compromise solution.

From a unit test, you can call `$this->skipMiddleware([])`, and pass a list of middleware to skip. For example:

```php
$this->skipMiddleware([\App\Http\Middleware\VerifyCsrfToken::class]);
$this->post('foo', ['bar' => 'baz']);
```

The most common use case would be wanting to ensure the  user is authorized to perform an action via middleware.