Alias sort functions on Collection
On eloquent we use `orderBy` - but on collection classes we use `sortBy`

It would be nice to alias the collections to allow `orderBy` etc. This makes the code simpler, since you can use the same syntax.

If this is accepted - maybe we could/should add an alias to the Eloquent classes for `sort` (as an alias for order)?

And if this is accepted I'll send a PR for the docs.