Bug fix in method has() - Request.php
Inversion of terms in ternary operator within method has()

Method was failing to match multiple keys passed by means of an array. However, it would appear to be working properly if argument were a string or an array with a single key. Reason: the ternary operator has its terms inverted. func_get_args() should be the result of the condition is_array($keys)