Notification UUID set prior to queuing to ensure it is consistent.
When a notification implements ShouldQueue, different queued jobs are
created for each channel, and as the UUID is set in the execution of
each job it will be different for each channel (despite relating to the
same notification).  Without ShouldQueue implemented, the UUID is the
same throughout all channel executions.

This commit changes that to set the UUID on the notification once prior
to queuing, and then accessing that id when executing the jobs on each
channel.

This means we can always rely on $this->id returning the same value
within the toMail, toDatabase, etc methods even when a notification is
queued.

**Use Case**
I need to know that an email received, and a DatabaseNotification
record were generated by the same Notification.  This way the email URL
can link them back to the database version of the notification in my
application - so that I can mark it is read before moving them on to
the actual content I want them to see.  Without this, a user can click
on a mail notification to link into my app, but there is no association
with the DatabaseNotification I can use to mark it as read - so they’ll
see the same notification in an unread state in the application.

I think this is potentially a broad use case, but even if not hopefully
it is appealing to have $this->id return the same value in each channel
on the Notification regardless of whether it was queued or executed
live. Thanks for the consideration! :)