Updates to configuration loader
There was an issue when loading configuration file in Linux/Unix. When file is actually a symbolic link, the path gets messed up and none of the configuration provided in the target file is accessible because the path with dot notation contains everything in between he link and the target file.