Allow control over the short property on slack attachment fields.
In order to give access to the **short** setting on a Slack attachment
field, *optionally* allow an array to be passed in.

Example configuration could then be:
```
->fields([
    'ID'     => 99,
    [
        'title' => 'Description',
        'value' => $description,
        'short' => false,
    ],
]);
```

This ensures that **description** will be nicely rendered on it’s own
line in Slack.  Without this, Slack will render it in 50% width even
when it is not necessary (mobile = always, desktop = if next to another
field) which does not look nice.