Keep array keys only if array is Associative
When you have a sequential array, and you sort it using sortBy,
you would expect to get a sequential array back. This is especially
an issue when you try to cast it to Json, and you get an object
instead of an array. This commit checks if an array is Associative,
and if not, discards the keys creating a new sequential array instead.