Change maintenance down file path
Change maintenance down file path so we can in the case you run laravel in a docker container mount the down file on the host by mounting the storage/framework/state folder. 
Use case : when doing a blue green deployment with docker container where the aim is to deploy new container version then switch container in the load balancer, the goal is to keep the state of the app. As there is still a bug with docker when trying to mount a single file as a volume if the file does not exists, and as the down file is in framework folder which has the cache, sessions and view folders, we cannot mount this folder. So having the down file in another folder allows us to mount this folder and keep the maintenance state between deployments.