Resolve sub-folders for local filesystem.
When working with local file system, the url method will only prepend the folder '/storage/' to the specified path of the file. In the event of the developer specifying multiple local storage and each one having it's own sub-folder in the public folder, the url() method will not correctly resolve the file even if a disk is specified.