Allow Proper Wrapping of Database Names Containing Spaces When Aliasing Columns (Perfectly Legal in SQL Server)
Currently, when using SQL Server, if you have a database which includes spaces, the grammar handles wrapping it fine. 

So this:

My Sample Database.dbo.my_table.my_column

Becomes:

[My Sample Database].[dbo].[my_table].[my_column]

But if you also try to alias a column when the database contains spaces, it breaks. 

This:

My Sample Database.dbo.my_table.my_column AS my_aliased_column

Becomes:

[My] AS [Sample]

I have traced the error to Database/Grammar using explode( ' ' ) - the code thinks the spaces in the database name are the spaces around the "AS".

You can fix this by using regex instead of explode, splitting on the "AS" more accurately.

Just replace the explode with this:

$segments = preg_split( '/\s+as\s+/i', $value );

Please note: When you have an Eloquent belongsToMany method that has to include the database name (eg. we need to join across DBs), this type of alias is created automatically, generating this error.