"trans" helper is not calling getFromJson method
"trans" helper method is not using new json file translation. Therefore, when you called "trans" helper method, json "sentence translation" is not working. I am not sure that this was made intentionally. If I use "__" helper method, there is no problem. So that is the difference between the two. Since latest 5.4 documentation is not covering "trans" helper method on localization page, any recommandation to use either of them?