[5.4] Remove unused "deleteToken" method
According to #1752: in commit c454c87, reset token will be hashed first before saving to the database. So it is not possible to delete the entry based on the public token sent to the user. Accordingly the `deleteToken` method can't be used this way, instead `$this->tokens->delete($user);` has to be used, which accepts user instance anyway.