Problem in paginate with distinct and joins
When using the distinct(), join() and paginate() methods the total with pagination is different from the total without pagination. The solution is to put the distinct field in the paginate method, for example paginate (10, ['events.id']), but the getCountForPagination function does not pass this parameter to get the total correct value.