[5.4] Allow lang path to be an array
My use case:

I'm working on a multi-tenant/multi-language website which needs to have
a global languages directory in 'storage/lang' and a local (domain
specific) languages in the usual 'resources/lang' directory.

The local domain specific messages overrites the global messages.

For now I just extend the \Illuminate\Foundation\Application and
override the langPath() method supplying an array of possible paths for
the \Illuminate\Translation\FileLoader.php to check.

However this can be further customized to allow for languages to be
loaded from a 'paths' key in the config/app.php (or even a separate
config/lang.php file) just like the config/view.php has an array of
possible paths.