Adds a feture to allow simple datetime formatting
A developer can easily list any data/datetime field in a property
called $formattedDates in the model. Each listed property will be
formatted each time it is accessed or set. The format can be specified in the
configuration file. Additionally, the developer can easily specify a
different format for each property.