Use the built-in IP Address type if supported
The ability to use the built-in IP Address type in databases that support it was added in f89c580353826fae429bc7c386a92e62a874a251 (v5.2.27). It probably should be used here instead of defaulting to a string. Any database that doesn't have this type built in will still fallback to a 45-character string field.