Allow for expiry without nx or xx set
In the old situation, it's not possible to set an expiry without setting $expireResolution (either to nx or xx or to an nonsense value). As using nx or xx is optional for redis and it's perfectly acceptable to use an expiry without setting nx or xx, it should also be possible to use $expireTTL and $flag without setting $expireResolution (or rather, when $expireResolution is null).