Implement `until` method on the `EventFake`
When dispatching events using the `until` method instead of the
`dispatch` or `fire` methods, the given events were not being collected
in the `events` array of the `EventFake` class.

This was resulting false negatives, e.g. `assertDispatched` is failing
because the `until` method is not implemented like the `fire` and
`dispatch` methods.

* Fixes #18923