Make sure that Carbon uses the correct locale
When using Laravel, you may want to serve a site in multiple languages.
Eloquent returns date objects that can be displayed in human relative
text form. We have a locale that is set by default when the app
boots, and it can also be called using App::setLocale('code')

Users expect objects returned from the framework to be in the locale
that they set in the configuration options.