Foreign helper: Relationship + dropRelationship
Typically, creating a relationship requires 3 methods: `foreign()`, `references()`, `on()`. With this helper, the user will only have to use a single method with 3 parameters (two required, one optional).

Similarly, to drop the relationships, you have to remember the default syntax (`TABLE`_`COLUMN`_foreign). With this helper, users won't have to remember.

`relationship()` still allows for chain with `onDelete` or similar.

Maybe this is a terrible idea? If so, please let me know :) It just feels easier and I imagine it would lover the barrier for people new to the whole migrations concept (as I remember having some trouble with the whole `foreign()->references()->on()`)