do not reuse variable name in foreach loop
due the scoping of PHP foreach loops, this works fine. however it can be a little confusing from a readability standpoint.

I just rename the value of the first loop so there is no question which variable we are referring to.

also, if we were to ever use the `$aliases` variable after this loop, it would not contain what we expected.