Add 404 request method and url when using withoutExceptionHandling
Whenever a NotFoundHTTPException is thrown, it is helpful to see the request method and url. Helps to debug what route to look at instead of having to first look it up in the test file. I've been using this in my own tests and I love it.

Final result:

```
1) Tests\Feature\Backstage\PublishConcertTest::a_promoter_can_publish_their_own_concert
Symfony\Component\HttpKernel\Exception\NotFoundHttpException: POST http://ticketbeast.app/backstage/published-concerts
```