Resolve double spending scheduled event
When having multiple server running the scheduler. If a job take only a couple of hundred of miliseconds to execute, the withoutOverlap won't stop a second server to take the same event.

Adding an artificial delay after job completion of a couple of second will resolve this.

A cron can only at least be launched every minute, the seconds added won't influence the normal working of the scheduler.