Add Collection::wrap method
If the given value is not a collection, wrap it in one.

If you pass a collection or an array as the value you get a collection with data in the collection or array back. If you pass in anything else you get a collection of one containing that value.