Allow up to 11 options to be presented for consistency.
Ensures that getSliderTooCloseToBeginning returns 9 pages in the first
property (working with an offset of 3), so that when combined with the
two in the last property, there are 11 clickable page buttons in the
rendered paginator.

This keeps it consistent with getSliderTooCloseToEnding which returns 2
options + ‘…’ + 9 more options (total 11), and with getFullSlider which
returns 2 options + ‘…’ + 7 options + ‘…’ + 2 options (total 11).