allow random to return 0 items
currently passing in a 0 to the `Collection->random()` method will result in a php error.

this is because it passes the parameter through to the native php `array_rand` function which specifies the 'second argument needs to be between 1 and the number of elements in the array'.

now if we pass in 0, we will catch that early, and simply return an empty static.