add ability to define reloaded relationships
currently when a Model is added to a queue and serialized, the
relationships of the Model are not restored when the Model is
unserialized/refetched from the database.

When you are handling the queued event/job with PHP, it’s not a big
issue. You can either manually load the relationships, or just use them
and they will be fetched when needed.

However, when you are broadcasting the events and listening for them in
Javascript, currently there is no way reload the relationship.  This
addition allows us to define which relationships we would like to be
reloaded when the object is unserialized.