Fix Auth command views stubs
passwords/reset.stub was handling a success status message stored in
session, but there isn't a flow that redirects to that view with said
message since the user is authenticated and redirected into the
application itself when the password is successfully changed. This
commit moves the status message from the passwords/reset.stub to the
home.stub so it's clear to the user that his password was changed once
he is logged in.