Fixed bug #20943  'before' method isn't called
The deleted lines is asking all abilities to be written into callable methods in policy classes.

e.g. in the policy class bellow:
```
class Policy
{
    public function before(User $user, $ability)
    {
        //
        dd('The \`before\`method is called.');
    }

    public function someAbility(User $user, Class $object)
    {
        //
    }
}
```

when we authorize in a controller `$this->authorize('someAbility', $someObject)`, we may get `The \`before\`method is called`; 
when we authorize in a controller `$this->authorize('otherAbility')`, we may get `AccessDeniedHttpException`.

This commit solves this problem.