remove extra call to `formatNotifiables()`
this is a minor optimization.

the `formatNotifiables()` method is also called in both the `sendNow()` and `queueNotification()` methods, which `send()` defers to.  therefore, there is no need to call it here.

I would have preferred to leave it here and remove it from the `sendNow()` and `queueNotification()` methods, but `sendNow()` is a public method, so it could possibly be skipped over if I did it that way.