Fix boolean changes on eloquent models
If an eloquent model has a boolean field, it will be 1 or 0 in the database. If the model is then updated (with either true or false), originalIsEquivalent() will return false because true !== 1. However, if $current on the eloquent-model is a boolean, we can safely assume that the field is boolean and we can therefore treat $original as a boolean as well.