set user provider for viaRequest guards
You may want to use the user provider in the callback:
```
Auth::viaRequest('api', function ($request, $provider) {
    return $provider->retrieveById(1);
});
```

My fix ensures that the provider specified in auth.api.provider is used.
The current behavior is to use auth.defaults.provider for all viaRequest
guards, but that configuration is not set by default such that $provider
is null.