Remove the integer key check in session errors comparison
In the `TestResponse::assertSessionHasErrors`, if the error key is an
integer, the error value is sent to the message bag to determine if any
message exists for it, which is incorrect because the message bag uses
the key to determine if any message exist for it.

With this changes, the numeric key can be used to fetch the error
messages and determine if the given value is contained in them.