Improve session errors assertions
In the previous implementation of `TestResponse::assertSessionHasErrors` i
found that was not possible to find errors that had an integer key.

This was happening because if the assertion detected that the key was an
integer, it was sending the error value to the `MessageBag::has` method,
which was incorrect because that method uses the key to determine if
there is any messages for it.