#23327 Moving code from the query builder to the eloquent builder
A previous pull request #19013, which aimed to solve the issue of multiple select statements causing an SQL error within a withCount function, assumed a sub query would only need one select statement.

This is not the case, so the code that removes excess filters is moved to the withCount method where the sub query does only need one select.