PHP 7.1 support for database migrations in test environment
In PHP 7.1, getNextBatchNumber generates a 'non-numeric value encountered' warning if the migrations table is empty, because getLastBatchNumber returns a non-integer, since the max method takes the max of nothing. In the test-environment, this warning becomes an error, and so all tests that use the database (an in-memory sqlite database in my case) automatically fail.