remove Blade defaults
current Blade functionality allows the user to define a default value if the first value is not set.

`{{ $name or 'Andrew' }}`

This feature was added in an early version of Laravel to be a shorthand for

`isset($name) ? $name : 'Andrew'`

PHP7 has introduced the null coalesce operator (`??`) which performs this exact functionality.

`$name ?? 'Andrew'`

While the 'or' feature was good when it was needed, it could produce unexpected results when the characters 'or' appeared anywhere in the Blade expression.  This was solved with extensive regex, but was still somewhat error prone.

Now that Laravel requires PHP 7.1, we can utilize the null coalesce operator and remove the 'or' feature. User upgrade process will be a find and replace, and the number of characters will be identical.

```
{{ $name or 'Andrew' }}
{{ $name ?? 'Andrew' }}
```