Allow specifying connection on scheduled job
Jobs can currently specify a queue to run on, but unlike normal job dispatching - jobs that are dispatched through the scheduler can not specify a connection. In fact, the ability to specify the queue is not documented either (which I will make another PR on).
This minor change allows specifying both a connection and a queue for scheduled jobs, which I'd deem to be important functionality as it's not at all uncommon to have more than one queue driver.