Check for null when calling array_key_exists
When a provider is misconfigured (as one, but not the only example), `Support\Arr::exists()` can be called with a `null` for the `$array` argument. This would cause the subsequent call to `array_key_exists` to fail. The problem with this is that in some cases, the root of the problem (a badly configured provider, for example) might be covered by this exception.