Add the whereNot operator
By adding this operator we can negate a complex query in a expressive way.
For example, the query pointed by @DrowningElysium:

`select * from table where b=true and NOT (e != 1 and f < 5 and (g != null and h >= 150))`

To be done currently in laravel query builder it's necessary to invert all
operators (`and` to `or`, `!=` to `==`, `<` to `>=`, `>=` to `<`). If we
need to do the negation dynamically we will need to maintain two querys,
increasing the possibility of bugs. For example:

```
DB::table('table')
  ->where('b', true)
  ->when($invert, function ($query) {
    return $query
      ->where('e', '!=', 1)
      ->where('f', '<', 5)
      ->where(function ($query) {
        return $query->whereNotNull('g')
          ->where('h', '>=', 150);
      });
   }, function ($query) {
    return $query
      ->where('e', '=', 1)
      ->orWhere('f', '>=', 5)
      ->orWhere(function ($query) {
        return $query->whereNull('g')
          ->orWhere('h', '<', 150);
      });
   });
```

With whereNot we can write this query as:

```
$subquery = function ($query) {
  return $query
    ->where('e', '!=', 1)
    ->where('f', '<', 5)
    ->where(function ($query) {
      return $query->whereNotNull('g')
        ->where('h', '>=', 150);
    });
};

DB::table('table')
  ->where('b', true)
  ->when($invert, function ($query) use ($subquery) {
      return $query->whereNot($subquery);
  }, function ($query) use ($subquery) {
      return $query->where($subquery);
  });

// or just

DB::table('table')
  ->where('b', true)
  ->{$invert ? 'where' : 'whereNot'}($subquery);
 ```

Which is more simple and legible.

We only allow calling whereNot in callback form: $query->whereNot($callback);
because allowing the use as column `whereNot($column)` can confuse the developer
since the behaviour of not() in some fields, like varchar, is different
from php !$string.

Resolve https://github.com/laravel/ideas/issues/708