Add "always" to complement "sometimes"
It is possible to add rules using this contract by passing a closure that only returns true to the `sometimes` method.  It doesn't seem very easy to read.  `validator->sometimes(always = true)`.  

The other (easier but less fun) option is to just make `addRules` part of the interface.  Some code seems to assume that the validator object that implements the contract will implement `addRules`.  This works against programming to contracts.