Give preference to sqlsrv driver over dblib when sqlsrv driver exists.
As commented on [#24492] this small change will help solve most compatibility issues on different environments when Sql Server is used.

The official Sql Server driver is now available for all major vendor platforms, this will help developers have the same experience using sql server with laravel/php on Windows and Unix based OSs.

If for some reason the sqlsrv driver is not installed it will fallback to dblib.