Add the onEachSide option to Builder's pagination, so you can customize LenghtAwarePaginator paginator's items count. E.g.: MyItem::paginate(10, ['*'], 'page', null, 1);                                                 ^ It's not really handy, but it's clean and not many people will complain. The main benefit is that having the default (3) items per page in the paginator may end with too many links in the paginator; that could be a problem in some layouts.
Keeping this argument sequence does not break any existing feature in the framework, and I also used 3 as the default onEarchSide parameter like before.

However this would be a breaking change since I'm modifying an existing method signature.

This is a feature request for the maintainers, if they know how to do it better, they're welcome.